Use server-side RLock/RUnlock
- Unify RLock/Lock and the unlock variants into a single type, namely
  DRWMutex.

- Allows utmost one RLock on a single DRWMutex instance. i.e, only one
  among many go-routines that share a reference to DRWMutex could
perform a RLock. For those who need to hold a two RLock(s) on the same
name, they could use two different DRWMutex objects to do that. The
book-keeping is done at the server, so all the rules of reader-writer
locks apply. In summary, this design choice doesn't reduce the set of
synchronization patterns that could be expressed using this package. It
simplifies the implementation of DRWMutex significantly.

- Remove parallelReaders unit-test for the above reason.