Wrapper type around DRWMutex to prevent starvation of writers.
Simple type to embed two DRWMutexes to (in similar fashion to sync.RWMutex) remove the preference for read locks over write locks (which could cause a write lock to starve).

Note that it requires two DRWMutexes instead of just a single one as in the regular case, so it is effectively twice as expensive. So only use this if really necessary.