Add support to remove stale locks using a maintenance loop (#22)
Add a maintenance mechanism to remove stale locks. Stale locks are detected by querying back to the original requestor for the 'golden' state of the lock (relying on the fact that the 'localhost' server will never have 'network' problems and thus always have the correct state). In case the original server cannot be reached because of a network issue, the status of the lock is unknown and resolution of the lock will try to be obtained in a next maintenance cycle.

Other improvements & additions:
- uid added to (r)lock message
- added node & rpcPath from originator to (r)lock message so it is known to which server to query back
- added new test cases for single & multiple stale locks, and single and multiple clients with locks that crash
- locks that do not include its own 'localhost' server are not treated as valid locks (released immediately)

Basic algorithm for Stale Lock detection
--------------------

The algorithm relies on the fact that the 'localhost' server of the client that is trying to get the lock is always actively participating in granting the lock. This is a logical assumption since there's no 'real' network traffic for this request  and in fact this [commit](https://github.com/fwessels/dsync/commit/8d47e02265c4170ab95e9470a5bb89fae1fbf576) now guarantees this to be the case.

So we can say that the status of the lock at this server is the 'golden' status for this lock. We require that the client and server are running in the same process, ie a crash will take both down.

For each lock server a maintenance loop is added that iterates over all locks that are active and that will perform a 'check' on a lock with a certain (configurable) age (say a minute, so short lived locks are never checked).

For each lock to be checked a call is made back to the server of the originating client to query whether the lock is still active or not (based on the combination of name and the `uid` of the original request). Three things can happen:
- the lock is still active --> not a stale lock
- the lock is no longer active --> we have detected a stale lock and can remove it
- the call fails (for whatever reason, can be either network failure, server down, something else, etc.): we cannot draw any conclusions so we will retry during a future maintenance loop

The following table summarizes the different situations that can occur:

| Client/Server  | Network      | Has Lock | Remark |
| -------------- | ------------ | -------- | ------ | 
| Running        | Connected    | No       | Lock can be requested provided quorum is obtained |
| Running        | Connected    | Yes      | Lock can be released |
| Crashed        | Connected    | No       | No side effects |
| Crashed        | Connected    | Yes      | Client/Server crashed while lock was active: see scenario 1 |
| Running        | Disconnected | No       | Lock request will always fail due to absence of quorum |
| Running        | Disconnected | Yes      | Client/Server has lock but are unable to deliver release message to quorum: see scenario 2 |
| Crashed        | Disconnected | No       | No side effects |
| Crashed        | Disconnected | Yes      | Client/Server crashed while lock was active and network is down: see scenario 3 |

Scenario 1: Both client and server go down while the client was holding a lock. The check call in the maintenance loop at other servers will fail as long as the client and server remain down. Once client and server are restarted, the lock is obviously lost at both the client and the server, so during a next iteration of the maintenance loop stale locks will be detected at the other servers.

Scenario 2: Client and server are still running and the client got a lock while the network was still up. Now it is disconnected from the network and it releases the lock so the unlock message is effectively only delivered at its own 'localhost' server. As long as the network remains disconnected, the call to detect stale locks in the maintenance loop at other servers will fail. Once the network becomes available again, the other servers will detect during their maintenance loops that the lock is no longer active and remove their stale locks.

Scenario 3: Client and server were running while locking a lock but then crashed and got disconnected from the network. As long as they are not restarted and the network connection is restored, the other servers will not be able to detect any stale locks. So both the client and server need to be restarted and the network connection needs to be restored in order to the stale locks to be detected.