Relax minimum number of nodes to 2
Now that we have relaxed the quorum for reads to N/2, it actually makes sense to lower the minimum number of nodes from 4 down to 2.

This means that for a 2 node system, both nodes need to be up in order to acquire a (write) lock. To obtain a read lock just a single node (which one doesn't matter) needs to be up.